backend/requirements.txt
# Install dependencies
Flask
flask-cors
pytesseract
Pillow
spacy
